Add scripts to build & publish our github actions, to 'bare tags' in the Khan/actions repo
Summary:
Here's the magic! This will allow us to reference e.g. `Khan/actions@filter-files-v0.0.1` in our workflows (as described in the readme).
I've also added basic workflows, one to check for a changeset, and one to automate releases when a release PR lands.

Test plan:
See the workflow work! I published check-for-changeset-v0.0.0 manually from the CLI to demonstrate.