Allow check-for-changeset to be configured to exclude files by extension (#43)
## Summary:

The check-for-changeset action is really useful in making sure we don't put up a PR without a `.changeset/*` file. But it often flags PRs that we can guarantee don't need a changeset (for example, a test-only PR or one that just touches Storybook stories). 

This PR adds the ability to filter out entire classes of files by extension (*.test.js or *.stories.js) for example. 

We will still need to configure each repo that uses this action to exclude the files by extension, but this opens the door for those configurations.

Issue: "none"

## Test plan:

Use the action in this PR in a repo (Khan/perseus, for example) and create a PR that contains changes only to files with extensions we configure to exclude.

Author: jeremywiebe

Reviewers: jaredly, jeremywiebe

Required Reviewers:

Approved By: jaredly, jaredly

Checks: ✅ gerald, ✅ Lints

Pull Request URL: https://github.com/Khan/actions/pull/43