Try to fix our publish script
It has been publishing our actions correctly, but then the `changesets/action` fails, and so the "github action" technically errors out (see https://github.com/Khan/actions/actions/runs/4982973373/jobs/8919388693)
It would be nice for that not to be the case.

The thing that's happening is the changesets/action tries to `git push origin --tags`, and that's failing for some reason. I'll try doing it already within our script, to see if any good will come of it.